images: Add padded logo image
The Linutronix logo should be shown in the README.md of various
github projects maintained by Linutronix. Since it is not possible
to really control the image dimensions, add a version of the logo
that uses padding for 2/3 of the image. This will allow the logo
to appear in a palatable size.

Signed-off-by: John Ogness <john.ogness@linutronix.de>
Reviewed-by: Wilfried Wessner <wilfried.wessner@linutronix.de>